moved random word generator to helper function
